refactor: moved macro code to separate files
